Moved integration tests into separate package
