+++ app/src/androidTest/java/integration/screen/PlantScreen.kt
@@ -1,11 +1,10 @@
-package com.google.samples.apps.sunflower.screen
+package integration.screen
 
 import android.accessibilityservice.AccessibilityService
 import android.content.Intent
 import androidx.test.InstrumentationRegistry
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions.click
-import androidx.test.espresso.assertion.ViewAssertions.doesNotExist
 import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.intent.Intents
 import androidx.test.espresso.intent.matcher.IntentMatchers
